+++ src/App.js
@@ -4,10 +4,12 @@ import { Route, Switch } from 'react-router-dom';
 import LoginPage from './pages/LoginPage';
 import FindIdPage from './pages/FindIdPage';
 import FindPasswordPage from './pages/FindPasswordPage';
+import SignupPage from './pages/SignupPage';
 function App() {
   return (
     <Switch>
       <Route path="/shop/account/signin" component={LoginPage} />
+      <Route path="/shop/account/signup" component={SignupPage} />
       <Route path="/shop/account/find_id" component={FindIdPage} />
       <Route path="/shop/account/find_pwd" component={FindPasswordPage} />
     </Switch>
