+++ src/pages/SignupPage.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import Signup from '../components/signup/Signup';
+
+const SignupPage = () => {
+  return <Signup />;
+};
+
+export default SignupPage;
